Add ECU MISRA baseline, expanded tests, and firmware cleanup
Adds the first ECU MISRA C baseline and expands host-side validation for first-party ECU firmware.

MISRA baseline / CI:
- Adds ECU MISRA-lite hygiene checking to CI.
- Documents MISRA migration scope and exclusions in MISRA_BASELINE.md.
- Scopes compliance checks to first-party ECU firmware while excluding HAL, CMSIS, FreeRTOS, CubeMX-generated files, and host-test infrastructure.
- Adds checks for BOMs, hidden/bidirectional Unicode, non-ASCII first-party firmware text, reserved header guards, unsafe libc calls, malformed float suffixes, empty parameter lists, non-static task entry functions, and other MISRA hygiene risks.
- Keeps this as a baseline migration step rather than claiming full certified MISRA compliance.

MISRA cleanup:
- Replaces reserved double-underscore header guards with project-safe guards.
- Makes empty function parameter lists explicit with void.
- Gives FreeRTOS task entry functions internal linkage.
- Makes CLI command handlers and command table file-local.
- Replaces atoi usage with checked strtol parsing.
- Replaces ret |= function_call style status accumulation with explicit status merging.
- Converts RTC/APPS function-like macros into static inline helpers.
- Adds unsigned suffixes to integer macros where appropriate.
- Cleans simple float literal suffix issues.
- Converts map() to float arithmetic instead of long double.
- Replaces strtok-based CLI parsing with bounded manual tokenization.
- Removes remaining first-party strlen usage from ISR-adjacent UART callback logic.
- Makes STM32 mutex attribute objects static const.
- Initializes RTC local structs before HAL calls.

Driver hardening:
- Adds null guards to board, CLI, dashboard, potentiometer, pressure sensor, STM32 wrapper, NTC, PWM, flow sensor, CAN, and MPU6050 paths where safe.
- Hardens MPU6050 init/read behavior with config validation, null checks, HAL status propagation, zero-scale protection, and correct signed big-endian raw sample conversion.
- Renames the misleading MPU6050 register label for 0x6B to the power-management register.
- Hardens PWM init by rejecting invalid channels, setting duty to zero before start, and propagating HAL_TIM_PWM_Start failures.
- Hardens flow sensor init/read by recording timer start failures and clamping impossible high_count > by rejecting invalid channels, setting duty to zero before start, and propagating HAL_TIM_PWM_Start failures.
- Hardens flow sensor init/read by recording timer start total_count captures.
- Hardens CAN transmit by rejecting invalid standard IDs above 0x7FF.
- Cleans potentiometer and pressure sensor helpers with explicit float math, fabsf, null-safe behavior, and clearer percent handling.
- Fixes poten_percent_to_hex scaling to map 0-100 percent linearly to 0-65535.

Expanded host/unit testing:
- Adds a dedicated driver test runner and host-side HAL/CMSIS stubs.
- Expands Makefile targets so make ci runs unit, driver, regression, and analyzer checks.
- Adds CI execution of the expanded driver tests.
- Adds coverage for map, poten, pressure_sensor, pwm, flow_sensor, mpu6050, canbus, cli, dashboard, ntc, rtc helpers, and stm32f767 wrapper behavior.
- Adds edge-case tests for MPU6050 failed reads, signed sample conversion, init failure propagation, invalid config rejection, PWM channel validation, CAN timeout/ID handling, flow sensor impossible captures, CLI token limits, UART error propagation, dashboard write errors, and ADC wrapper guards.

Logic/detail cleanup:
- Fixes CLI tokenization capacity handling.
- Handles empty or whitespace-only CLI input safely.
- Ensures driver init/read functions reject invalid pointers or invalid configuration instead of silently succeeding.
- Updates firmware review notes to record the MISRA baseline, expanded testing, remaining coolant calibration note, and known scope limitations.

Validation performed:
- Project structure check passed.
- Repo hygiene check passed.
- ECU MISRA-lite hygiene check passed.
- git diff --check passed.
- ECU unit tests passed.
- ECU driver tests passed.
- ECU host regression tests passed.
- GCC analyzer passed.
- Repo hygiene remained clean after test cleanup.

The STM32 ARM-GCC headless firmware build should still be verified in GitHub CI or on a machine with arm-none-eabi-gcc installed before merging.

+++ Core/Src/ext_drivers/ams.c
@@ -31,12 +31,12 @@ void segment_init(segment_t *dev)
         return;
     }
 
-    for(int i = 0; i < NVOLTS; i++)
+    for(uint16_t i = 0u; i < NVOLTS; i++)
     {
         dev->volts[i] = 0u;
     }
 
-    for(int i = 0; i < NTEMPS; i++)
+    for(uint16_t i = 0u; i < NTEMPS; i++)
     {
         dev->temps[i] = 0u;
     }
@@ -51,7 +51,7 @@ void ams_init(ams_t *dev)
 
     memset(dev, 0, sizeof(*dev));
 
-    for(int i = 0; i < NSEGS; i++)
+    for(uint16_t i = 0u; i < NSEGS; i++)
     {
         segment_init(&dev->segs[i]);
     }
